Reorganize features/ to Apps, Blog, Portfolio
- Rename features/Toys → features/Apps; files become AppsIndex/AppPage/
  AppLayout/appsData (exports apps/appsByCategory/getApp). No 'Apps.jsx' —
  avoids confusion with the root App.jsx.
- Move ChineseZodiac into Apps/ChineseZodiac.
- Move Curiosities under Apps, with Themes and LogoSizer flattened inside it,
  so the only top-level feature folders are Apps, Blog, Portfolio.
- Keep tools/, toykit, and toy_* analytics event names unchanged.
- Update route imports and gen-seo source path.
