+++ src/features/Apps/ChineseZodiac/AnimalInfo.jsx
