+++ src/features/Apps/Curiosities/Curiosities.jsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom'
-import Container from '../../components/Container'
-import { trackEasterEgg } from '../../utils/analytics'
-import { useSeo } from '../../utils/seo'
+import Container from '../../../components/Container'
+import { trackEasterEgg } from '../../../utils/analytics'
+import { useSeo } from '../../../utils/seo'
 
 // A quiet hub for the off-the-main-path corners of the site, reached from the
 // footer easter egg. Each card routes onward to its own page.
