+++ src/features/Apps/tools/ClaudeVision.jsx
