+++ src/features/Apps/tools/InflationCalculator.jsx
