+++ src/features/Apps/tools/UuidGenerator.jsx
