+++ src/features/Apps/tools/WordCounter.jsx
